backend/requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
openai
google-generativeai
edge-tts
faster-whisper
